+++ ionoscloud/resource_application_loadbalancer.go
@@ -12,6 +12,8 @@ import (
 	ionoscloud "github.com/ionos-cloud/sdk-go/v6"
 	"github.com/ionos-cloud/terraform-provider-ionoscloud/v6/services"
 	"github.com/ionos-cloud/terraform-provider-ionoscloud/v6/services/cloudapi"
+	cloudapiflowlog "github.com/ionos-cloud/terraform-provider-ionoscloud/v6/services/cloudapi/flowlog"
+	"github.com/ionos-cloud/terraform-provider-ionoscloud/v6/utils"
 )
 
 func resourceApplicationLoadBalancer() *schema.Resource {
@@ -61,6 +63,16 @@ func resourceApplicationLoadBalancer() *schema.Resource {
 				Type:     schema.TypeString,
 				Required: true,
 				ForceNew: true,
+			},
+			"flowlog": {
+				Type:     schema.TypeList,
+				Optional: true,
+				Elem:     cloudapiflowlog.FlowlogSchemaResource,
+				MaxItems: 1,
+				Description: `Only 1 flow log can be configured. Only the name field can change as part of an update. Flow logs holistically capture network information such as source and destination 
+IP addresses, source and destination ports, number of packets, amount of bytes, 
+the start and end time of the recording, and the type of protocol – 
+and log the extent to which your instances are being accessed.`,
 			},
 		},
 		Timeouts: &resourceDefaultTimeouts,
@@ -147,6 +159,25 @@ func resourceApplicationLoadBalancerCreate(ctx context.Context, d *schema.Resour
 		diags := diag.FromErr(errState)
 		return diags
 	}
+	if flowLogs, ok := d.GetOk("flowlog"); ok {
+		fw := cloudapiflowlog.Service{
+			D:      d,
+			Client: client,
+		}
+		if flowLogList, ok := flowLogs.([]any); ok {
+			for _, flowLogData := range flowLogList {
+				if flowLogMap, ok := flowLogData.(map[string]interface{}); ok {
+					flowLog := cloudapiflowlog.GetFlowlogFromMap(flowLogMap)
+					err := fw.CreateOrPatchForALB(ctx, dcId, d.Id(), "", flowLog)
+					if err != nil {
+						d.SetId("")
+						diags := diag.FromErr(fmt.Errorf("error creating flowlog for application loadbalancer: %w, %s", err, responseBody(apiResponse)))
+						return diags
+					}
+				}
+			}
+		}
+	}
 
 	return resourceApplicationLoadBalancerRead(ctx, d, meta)
 }
@@ -168,11 +199,20 @@ func resourceApplicationLoadBalancerRead(ctx context.Context, d *schema.Resource
 	}
 
 	log.Printf("[INFO] Successfully retreived application loadbalancer %s: %+v", d.Id(), applicationLoadBalancer)
-
-	if err := setApplicationLoadBalancerData(d, &applicationLoadBalancer); err != nil {
-		return diag.FromErr(err)
+	fw := cloudapiflowlog.Service{
+		Client: client,
+		Meta:   meta,
+		D:      d,
 	}
-	return nil
+	flowLog, apiResponse, err := fw.GetFlowLogForALB(ctx, dcId, *applicationLoadBalancer.Id, 1)
+	if err != nil {
+		if !apiResponse.HttpNotFound() {
+			diags := diag.FromErr(fmt.Errorf("error finding flowlog for application loadbalancer: %w, %s", err, responseBody(apiResponse)))
+			return diags
+		}
+	}
+
+	return diag.FromErr(setApplicationLoadBalancerData(d, &applicationLoadBalancer, flowLog))
 }
 
 func resourceApplicationLoadBalancerUpdate(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
@@ -227,6 +267,34 @@ func resourceApplicationLoadBalancerUpdate(ctx context.Context, d *schema.Resour
 		request.Properties.LbPrivateIps = &privateIps
 	}
 
+	if d.HasChange("flowlog") {
+		old, newV := d.GetChange("flowlog")
+		var firstFlowLogId = ""
+		if old != nil && len(old.([]any)) > 0 {
+			firstFlowLogId = old.([]any)[0].(map[string]any)["id"].(string)
+		}
+
+		if newV.([]any) != nil && len(newV.([]any)) > 0 {
+			for _, val := range newV.([]any) {
+				if flowLogMap, ok := val.(map[string]any); ok {
+					flowLog := cloudapiflowlog.GetFlowlogFromMap(flowLogMap)
+					fw := cloudapiflowlog.Service{
+						D:      d,
+						Client: client,
+					}
+					err := fw.CreateOrPatchForALB(ctx, dcId, d.Id(), firstFlowLogId, flowLog)
+					if err != nil {
+						//if we have a create that failed, we do not want to save in state
+						// saving in state would mean a diff that would force a re-create
+						if firstFlowLogId == "" {
+							_ = d.Set("flowlog", nil)
+						}
+						return diag.FromErr(err)
+					}
+				}
+			}
+		}
+	}
 	_, apiResponse, err := client.ApplicationLoadBalancersApi.DatacentersApplicationloadbalancersPatch(ctx, dcId, d.Id()).ApplicationLoadBalancerProperties(*request.Properties).Execute()
 	logApiRequestTime(apiResponse)
 
@@ -295,13 +363,23 @@ func resourceApplicationLoadBalancerImport(ctx context.Context, d *schema.Resour
 	if err := d.Set("datacenter_id", datacenterId); err != nil {
 		return nil, fmt.Errorf("error while setting datacenter_id property for alb %q: %w", albId, err)
 	}
-
-	if err := setApplicationLoadBalancerData(d, &alb); err != nil {
+	fw := cloudapiflowlog.Service{
+		Client: client,
+		Meta:   meta,
+		D:      d,
+	}
+	flowLog, apiResponse, err := fw.GetFlowLogForALB(ctx, datacenterId, d.Id(), 0)
+	if err != nil {
+		if !apiResponse.HttpNotFound() {
+			return nil, fmt.Errorf("error finding flowlog for application loadbalancer: %w, %s", err, responseBody(apiResponse))
+		}
+	}
+	if err := setApplicationLoadBalancerData(d, &alb, flowLog); err != nil {
 		return nil, err
 	}
 	return []*schema.ResourceData{d}, nil
 }
-func setApplicationLoadBalancerData(d *schema.ResourceData, applicationLoadBalancer *ionoscloud.ApplicationLoadBalancer) error {
+func setApplicationLoadBalancerData(d *schema.ResourceData, applicationLoadBalancer *ionoscloud.ApplicationLoadBalancer, flowLog *ionoscloud.FlowLog) error {
 
 	if applicationLoadBalancer.Id != nil {
 		d.SetId(*applicationLoadBalancer.Id)
@@ -341,6 +419,19 @@ func setApplicationLoadBalancerData(d *schema.ResourceData, applicationLoadBalan
 			if err != nil {
 				return fmt.Errorf("error while setting lb_private_ips property for application loadbalancer %s: %w", d.Id(), err)
 			}
+		}
+	}
+
+	if flowLog != nil {
+		var flowlogs []map[string]any
+		result, err := utils.DecodeStructToMap(flowLog.Properties)
+		if err != nil {
+			return err
+		}
+		result["id"] = *flowLog.Id
+		flowlogs = append(flowlogs, result)
+		if err := d.Set("flowlog", flowlogs); err != nil {
+			return fmt.Errorf("error setting flowlog %w", err)
 		}
 	}
 	return nil